GANs/requirements.txt
absl-py==0.11.0
appdirs==1.4.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
dataclasses==0.6
decorator==4.4.2
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata==2.0.0
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1597876320876/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
pandas @ file:///tmp/build/80754af9/pandas_1602088128026/work
Pillow @ file:///tmp/build/80754af9/pillow_1603822253009/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658308664/work
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376882706/work
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
six==1.15.0
tb-nightly==2.4.0a20201101
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2020.10.1
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.7.0
torchvision==0.4.0a0+9232c4a
tornado==6.0.4
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
zipp==3.4.0
